Add samba_role: member
With "samba_role: member" a Samba Domain Member Server can be set up.

We only install the needed packages to make the server work. Espacially
we do not install 'libpam-winbind', 'libnss-winbind' and 'libpam-krb5'.
They are only needed if Linux user login directly to the machine (a so
called Domain Member Client).

If "samba_role: member" is set, 'security' and 'realm' in smb.conf are set
automatically. 'default_realm' is set in /etc.krb5.conf.

Also added a few examples in `defaults/main.yml` to make it easier for
the user to find the correct domain member settings (we actually only
used these settings).