Molecule: Rename scenario
In 41fa9ce I talked about renaming the 'default' scenario to 'member'
but actually did not do so. Did so now and created a 'default' symlink
to 'member'. That way 'molecule test' can run without specifying a the
scenario.

The github-workflow molecule fails because it can not find the ansible role in
the given directory (because the I have named the folder 'ansible-role-samba'
locally and the github repository is called 'ansible-samba' which doesn't
work). I now use 'luisico.samba' (namespace.name from meta/main.yml) which works.

The github-workflow linter also found some problems, but I only fixed the indentation
of 'tasks/member.yml'. I will report the other upstream.