Add dnsutils + ldb-tools + libpam-winbind
Although we are not planning to log on to the Linux server (as a domain
member client), the documentation says to install libpam-winbind, so we
install it too.

Strictly speaking, dnsutils + ldb-tools are not required to run a Samba
domain controller or member, but they are helpful tools and should be
installed by default.