Add option to enable Michael Tokarev's Samba Repository
To enabled Michael Tokarev's Samba Repository, `samba_mjt_repository: true`
needs to be set. If enabled the signing key is, the apt repository and
Samba 4.19 is going to be installed.
See:

- http://www.corpit.ru/mjt/packages/samba/
- http://samba.bigbird.es/doku.php?id=samba:installing-from-mt-repos

rsync is needed to replicate the SysVol folder. See:

- http://samba.bigbird.es/doku.php?id=samba:sync-sysvol

krb5-config is not really needed, but is mentioned in the tutorials.

Added the Ansible Molecule scenario 'dc-mjt' to where
`samba_mjt_repository: true` is set.

Updated the Github Action molecule.yml to test all scenarios.